+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import model.User;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.storage.FriendStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.util.List;
@@ -12,6 +13,7 @@ import java.util.Set;
 @RequiredArgsConstructor
 public class UserService {
     private final UserStorage userStorage;
+    private final FriendStorage friendStorage;
 
 
     public User create(User user) {
@@ -35,18 +37,18 @@ public class UserService {
     }
 
     public void addFriend(Long id, Long friendId) {
-        userStorage.addFriend(id, friendId);
+        friendStorage.addFriend(id, friendId);
     }
 
     public void deleteFriend(Long id, Long friendId) {
-        userStorage.deleteFriend(id, friendId);
+        friendStorage.deleteFriend(id, friendId);
     }
 
     public Set<User> getAllFriends(Long id) {
-        return userStorage.getAllFriends(id);
+        return friendStorage.getAllFriends(id);
     }
 
     public Set<User> getMutualFriends(Long id, Long friendId) {
-        return userStorage.getMutualFriends(id, friendId);
+        return friendStorage.getMutualFriends(id, friendId);
     }
 }
