+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -14,15 +14,6 @@ public interface FilmStorage {
 
     List<Film> getAll();
 
-    void addLike(Long id, Long userId);
-
-    void deleteLike(Long filmId, Long userId);
-
-    List<Film> getPopularFilm(Integer count);
-
     Film findFilmById(Long id);
-
-//    List<Film> get();
-
 
 }
