+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #server.port=8080
 spring.sql.init.mode=always
 # ? jdbc-url ???????, ??? ?????? ????? ????????? ? ????
-spring.datasource.url=jdbc:h2:file:./db/filmorate
+spring.datasource.url=jdbc:h2:mem:./db/filmorate
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=password
